refactor: move health trend contracts to output crate
